Replace Pipenv with Pip
This is a middleware which could work with multiple versions of Python,
Django and different types of OS environments. This is better reflected
with Pip which doesn't pin down the version of Python or the libraries
and is the motivation behind moving to Pip.

Pipenv ensures that the virtual environment is exactly the same as
specific in the Pipfile.lock file down to Python version and library
verions. This does not reflect where this middleware is expected to run.